Refactor: move ComboBoxWithSearch to controls package
