+++ desktop/src/main/java/bisq/desktop/main/content/bisq_easy/offerbook/BisqEasyOfferbookView.java
@@ -22,8 +22,8 @@ import bisq.desktop.common.Layout;
 import bisq.desktop.components.containers.Spacer;
 import bisq.desktop.components.controls.BisqIconButton;
 import bisq.desktop.components.controls.BisqTooltip;
+import bisq.desktop.components.controls.ComboBoxWithSearch;
 import bisq.desktop.components.controls.SearchBox;
-import bisq.desktop.components.overlay.ComboBoxWithSearch;
 import bisq.desktop.main.content.chat.ChatView;
 import bisq.i18n.Res;
 import de.jensd.fx.fontawesome.AwesomeIcon;
