+++ desktop/src/main/java/bisq/desktop/main/top/MarketPriceComponent.java
@@ -27,8 +27,8 @@ import bisq.desktop.common.threading.UIScheduler;
 import bisq.desktop.common.threading.UIThread;
 import bisq.desktop.common.utils.ImageUtil;
 import bisq.desktop.components.controls.BisqTooltip;
+import bisq.desktop.components.controls.ComboBoxWithSearch;
 import bisq.desktop.components.controls.ProgressBarWithLabel;
-import bisq.desktop.components.overlay.ComboBoxWithSearch;
 import bisq.i18n.Res;
 import bisq.presentation.formatters.DateFormatter;
 import bisq.presentation.formatters.PriceFormatter;
